Integrate AbstractEventStoreConcurrencyTestBase into existing AbstractEventStoreTestBase
